New: Put image file path in the error messages
One of the difficulties in addressing errors and crashes in IIP is tracking down which images are causing a problem with reading and writing. On production systems, log levels are usually set to ERR, but this log level gives very little information about the current image being processed.

This commit adds the IIPImage path to the error message so that any error messages include a pointer back to the original image that caused the problem.